Move rate limiter to modules
